[iOS][SDK] Setup the Authentication logic (#17752)
# Pull Request

- Related to: #17553

## Summary

This PR creates the foundation for the authentication with the ability to `authenticate`/`logOut`/`getAccessToken`.

Customers are responsible for providing the first access & refresh tokens. They must do so by communicating between their server and our. We do not persist them, but we manage the automatic refresh for our customers while the app is alive.

The authentication is the following:
- customers call `NablaClient.shared.authenticate` with some `NablaAuthenticationProvider`.
- the `NablaAuthenticationProvider` is called immediately to provide the first access & refresh tokens.
- from then, `Authenticator` will refresh those tokens if the `accessToken` expires.
- if the `refreshToken` expires, we will call the `NablaAuthenticationProvider` again.

The actual http request to do the token refresh is not implemented yet, waiting on `HttpManager` PR.