[iOS][SDK] GQLClient (#17956)
# Pull Request

- Fixes: #17696

## Summary

Introduces the `GQLClient` with `fetch`/`perform`/`watch` capabilities. It introduces a wrapper for the `ApolloClient`, `ApolloStore` and the different transport types. While they are not all needed yet. It's better to have them now to add features later (like ability to know if the websocket connection is lost on the `WebsocketTransport`).

I've also imported the ability to automatically re-trigger watchers when the network comes back, or when the app enters foreground again.

I'll create new issues for:
- apollo errors handling
- plug `Authentication` with the `GQLClient` (backend PR needed)
- enable `subscriptions` (backend PR needed)