[iOS SDK] Generic query pagination (#18476)
# Pull Request

- Related to: #18208

## Summary

I recommend commit per commit review.

This PR introduces `GQLPaginatedWatcher` as a generic class to handle the management of cursor based queries that manually update the cache to append items.

I also used it on the `watchConversationItems` endpoint but I am waiting for a fix on the backend before implementing the load more UX in the chat screen.

## Test plan

Create many conversations. Observe that pagination on conversation list works.


https://user-images.githubusercontent.com/2571118/164676748-b3ef97b6-762b-4056-8d8f-a43af0589bef.mov


# Pull Request

- Related to: #18208

## Summary

I recommend commit per commit review.

This PR introduces `GQLPaginatedWatcher` as a generic class to handle the management of cursor based queries that manually update the cache to append items.

I also used it on the `watchConversationItems` endpoint but I am waiting for a fix on the backend before implementing the load more UX in the chat screen.

## Test plan

Create many conversations. Observe that pagination on conversation list works.


https://user-images.githubusercontent.com/2571118/164676748-b3ef97b6-762b-4056-8d8f-a43af0589bef.mov