[iOS SDK] Fix conversation list retry (#19037)
## Summary

The current code tries to load more items when the user hits `Retry`. But it should refetch the conversations list instead.

The result is that when you tap "Retry" you stay in the loading view for ever. Because loading more items can not set you back to the `loaded` state.

This PR does it by re-wiring the `watcher`.

## Test plan

- Turn backend off
- Tap "Retry"
- Observe very quick loading but back to error view (instead of inifinit loading)
- Turn backend on
- Tap "Retry"
- List appears