[iOS SDK] Unit tests: ConversationItemsTransformer (#19211)
# Pull Request

- Related to: #18775

## Summary

Extracts the logic of transforming the `ViewItems` into a single and testable `ConversationItemsTransformer` class.

It appears that date separators with multiple groups of messages is broken. It creates duplicates. I've added a test that demonstrate the issue but won't have time to fix now. I've added a TODO for @ThibaultTourailles since you wrote that code. I also suspect the same issue to exist on the B2C app (there is no visual impact because we filter out duplicates in the views).

I've introduced `XCTRequire` methods as part of this PR. They behave like their `XCTAssert` equivalent but stop the test execution when they fail. It avoids adding extra `guard let` to prevent index out of bounds or casting issues during tests.

## Test plan

- CI