[iOS SDK] Check provide tokens and directly renew if needed (#19649)
# Pull Request

- Fixes: #19638

## Summary

This issue in #19638 is that manually generated tokens have expired when returned from `SessionTokenProvider.provideTokens`. While this scenario make no sense in production, it is not a great experience for the sample app.

The fix is to check the expiration of the provided tokens and renew them directly if need be. If the `refreshToken` is also expired, we return an explicit `authenticationProviderDidProvideExpiredTokens` error.

### Changelog

- Automatically refreshes expired `accessToken` returned by `SessionTokenProvider.provideTokens`.

## Test plan

- Repro steps from #19638