[iOS SDK] Address integration tests flakyness (#24426)
- Related to: #22332

## Summary

Integration tests are flaky currently, sometime failing due to "missing
request in the cassette".

I've spent time investigating the issue and believe I've found the route
cause:

Integration should NEVER call the endpoint to refresh the session's
tokens. To ensure this, we actually set up tokens that expire in the far
future when we record tests. But such request is the one missing in the
cassette.
There are only 2 ways to call `fetchTokens(refreshToken: Token)` and
cause the request that should not happen:
- from `requireTokens(session: Session)` (which is the path where we ask
the `SessionTokenProvider` to provide tokens), if they provide us
expired ones. This is mocked during integrations tests, and ALWAYS
returns tokens expiring in the far future.
- directly from `renewSession(_ session: Session)` (which is the general
path) and will call `fetchTokens(refreshToken:)` if the `refreshToken`
is not expired, but the `accessToken` is (otherwise `getAccessToken`
would have returned it directly).

Option 1 is safe we control what the `SessionTokenProvider` returns with
a stub in `TestEnvironment`, the issue must come from option 2.

For option 2 to be the issue, we need the `accessToken` to be `nil` or
expired but `refreshToken` to be non-nil and not expired. Given that we
provide fake tokens that expire around year 2054 for both, that's weird.
One thing to notice tho, is that we don't check the value of
`accessToken` and `refreshToken` at the same time. And we use `Session`
to keep a reference to them, `Session` being a `class`.

After many many many retries, I've managed to reproduce a race condition
where we call `getAccessToken` while the `session.token` is nil. We then
decide to make a new task to renew it. But by the time we check if the
`refreshToken` is available, `session.tokens` gets updated and we have a
valid `refreshToken` available to make a call to `fetchTokens`. Hence
causing a request that our cassette does not know about, and failing
tests.

The proposed solution here is to make `Session` a value type semantic.
Hence the `AuthenticatorImpl` will make a copy of its value when
`getAccessToken` is called, and the whole flow will use immutable tokens
until it's done.

With this change in place and in case of the exact same race condition,
`refreshToken` will remain `nil` in the copy and the `TaskHolder` will
return the shared task that already completed. Immediately returning the
correct access & refresh tokens.

## Test plan

- Rerun integration test ~20 times on local machine.
- Rerun CI ~10 times (for some reasons, the CI is more likely to
reproduce it than my machine, and CI takes longer).