[iOS SDK] ViewModel on the MainActor (#24382)
- Related to: #20516 & async/await refactor

## Summary

I keep doing more training on the new structured concurrency for swift.
And our `_viewModel.onChange` pattern in UIKit views is keeping us safe
from background thread changes in the VM, I think it's best to follow
the official recommendation (designed for SwiftUI).

This PR applies the `@MainActor` on our VMs. A VM should dispatch work
with `async` methods, but only mutate it's state on the main thread. The
`@MainActor` enforces it.

We could remove the `nonisolated` from our VMs `init` methods, but that
would expose a `@MainActor` on the `NablaSchedulingViewFactory`, which I
believe will trigger some customers for not much benefits.

I'v done it for the `AppointmentCellViewModel` tho, to be able to
perform an first round of update in-sync, and because it's a private
class.

### Changelog

None.

## Test plan

QA Scheduling.